+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [7.4.1 - 2021-10-18]
 ### Fixed
 - Quality Report for large amount of data
+- QB-Generator: Cache number of patients per mdrid (and per mdr id and value)
 
 ## [7.4.0 - 2021-09-13]
 ### Added